fix: 팀스페이스 입장 dto 분리
