+++ src/main/java/SSOP/ssop/dto/EnterTeamSpDto.java
@@ -1,0 +1,13 @@
+package SSOP.ssop.dto;
+
+public class EnterTeamSpDto {
+    private int inviteCode;
+
+    public int getInviteCode() {
+        return inviteCode;
+    }
+
+    public void setInviteCode(int inviteCode) {
+        this.inviteCode = inviteCode;
+    }
+}
